requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
flask==3.0.0
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
cloudscraper==1.2.71
selenium==4.15.2
webdriver-manager==4.0.1
redis==5.0.1
celery==5.3.4
